Move metrics tracker into plugin reporting
Relocated createMetricsTracker from src/shared/metrics to src/plugin/src/reporting so the plugin-only helper lives with its reporting components. Updated plugin imports/tests to use the new barrel and removed the redundant shared shim.
