--- src/shared/metrics.js
@@ -1,1 +1,0 @@
-export * from "./metrics/metrics-tracker.js";
